Fix versioning
Replace import for __version__ with text parsing.  This eliminates
problems with installing without Flask already there.  There is a
good discussion about this on stackoverflow, but the best solution
there really didn't work work for a single file package.  There is
likely a better way.  When it appears, I will fix it.

http://stackoverflow.com/questions/2058802/how-can-i-get-the-version-defined-in-setup-py-setuptools-in-my-package